formatted emeritus section
