+++ src/pages/community/meshmates.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import { ThemeProvider } from "styled-components";
 import theme from "../../theme/blog/themeStyles";
 import { GlobalStyle } from "../../sections/app.style";
@@ -6,8 +6,7 @@ import SEO from "../../components/seo";
 import Navigation from "../../sections/Navigation";
 import Footer from "../../sections/Footer";
 import Layout from "../../components/layout";
-import Meshmates from "../../sections/Meshmates-grid/Meshmates";
-
+import MeshmatesGrid from "../../sections/Meshmates-grid";
 
 const MeshmatePage=() => {
     return (
@@ -16,12 +15,11 @@ const MeshmatePage=() => {
                 <GlobalStyle />
                 <SEO title="Meshmates | Layer5 - The Service Mesh Company" />
                 <Navigation />
-                <Meshmates />
+                <MeshmatesGrid />
                 <Footer />
             </Layout>
         </ThemeProvider>
     );
 }
-
 
 export default MeshmatePage;
