+++ src/sections/Emeritus-grid/emeritus.style.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const EmeritusWrapper=styled.div`
+.profile-card{
+    width: 170px;
+
+    header {
+        padding: 15px 0 0;
+
+        h1 {
+            font-weight: 200;
+            font-size: 20px;
+            letter-spacing: 0px;
+        }
+        
+        
+    }
+}
+`;
