+++ src/sections/Meshmates-grid/index.js
@@ -6,9 +6,42 @@ import PageHeader from "../../reusecore/PageHeader";
 import ProfileCard from "../../components/Profile-card";
 import { MembersGridWrapper } from "../../sections/Members-grid/membersGrid.style";
 import ProgramDetails from "./ProgramDetails";
-import Emeritus from "../Emeritus-grid/Emeritus";
+import EmeritusGrid from "../Emeritus-grid";
+import { useStaticQuery, graphql } from "gatsby";
 
-const MeshmatesGrid=props => {
+const MeshmatesGrid=() => {
+    const data=useStaticQuery(
+        graphql`
+            query meshmates {
+                allMdx(
+                    sort: { fields: [frontmatter___name], order: ASC }
+                    filter: { fields: { collection: { eq: "members" } }, frontmatter: { meshmate: { eq: "yes" } } }
+                ) {
+                    nodes {
+                        id
+                        frontmatter {
+                            name
+                            github
+                            twitter
+                            status
+                            image_path{
+                                childImageSharp{
+                                    fluid(maxWidth: 200){
+                                        ...GatsbyImageSharpFluid
+                                    }
+                                }
+                                extension
+                                publicURL
+                            }
+                        }
+                        fields{
+                            slug
+                        }
+                    }
+                }
+            }
+        `
+    );
     return (
         <MembersGridWrapper>
             <PageHeader title="Community Meshmates"
@@ -21,7 +54,7 @@ const MeshmatesGrid=props => {
                             Layer5 MeshMates are committed to helping community members be successful contributors. MeshMates aid in identifying areas of projects to engage within, working groups to join, and in helping community members grow in their open source and cloud native knowledge. By connecting one-on-one, MeshMates will share tips on how to have the best community experience possible.
 </p>
                         <Row>
-                            {props.data.allMdx.nodes.map(({ id, frontmatter, fields }) => (
+                            {data.allMdx.nodes.map(({ id, frontmatter, fields }) => (
                                 <Col xs={12} sm={6} lg={4} key={id}>
                                     <Link to={fields.slug}>
                                         <ProfileCard frontmatter={frontmatter} />
@@ -34,7 +67,7 @@ const MeshmatesGrid=props => {
 </p>
                     </div>
                     <ProgramDetails />
-                    <Emeritus />
+                    <EmeritusGrid />
 
                 </Container>
             </div>
